02-use-cases/AWS-documentation-assistant/pyproject.toml
[project]
name = "aws-docs-agent"
version = "0.1.0"
description = "AWS Docs agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.1",
    "bedrock-agentcore-starter-toolkit>=0.1.2",
    "strands-agents>=1.2.0",
    "strands-agents-tools>=0.2.3",
    "uv>=0.8.4",
]
